Common Anki export tooling placed in mixin
