+++ ruby/lib/goi/pipeline/core.rb
@@ -6,3 +6,4 @@ require_relative 'core/exporter'
 require_relative 'core/pipeline'
 
 require_relative 'core/safe_processor'
+require_relative 'core/anki_exportable'
