+++ ruby/sig/lib/goi/pipeline/core/anki_exportable.rbs
@@ -1,0 +1,25 @@
+module Goi
+  module Pipeline
+    module Core
+      # Mixin to add common tooling around Anki output handling
+      module AnkiExportable
+
+        def write_file_headers: (IO io) -> void
+
+        def deck: () -> String?
+
+        def note_type: () -> String?
+
+        def tags_column_index: () -> Integer?
+
+        def header_row: () -> [String]?
+
+        def to_array_field: ([untyped] array) -> String?
+
+        def humanize_const: (String s) -> String
+
+        def tagize: (String s) -> String
+      end
+    end
+  end
+end
